CatAnimApplied
I am doing this in two just so that way if there is a problem with the scene I can redo that part.
Animation applied to player prefab
There is a problem with the player controller in that when the player runs off (not jump) the player can still jump in the air.